requirements.txt
telethon==1.34.0
SQLAlchemy[asyncio]==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-telegram-bot==20.8
openai==1.23.2
